requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
python-decouple
django-rest-framework
celery